Set up views of nektar with findable cmake configs
For some reason, cmake can't seem to find the config files for
nektar++ where it defaults to putting them. I'm not sure why, as the
documentation suggests it should be able to. In any case, in a spack
build environment this can be worked around by modifying
CMAKE_PREFIX_PATH. This option is not available for views, so instead
I added a symlink where cmake can find it.