Add NESO package to repo (#3)
* Removed unnecessary run-dependency settings for Nektar++

* Relaxed version requirements for hipsycl

This allows it to use consistent BOOST versions as Nektar and NESO

* All dependencies to find cmake configs for nektar

There was an attempt to set the CMake prefix path, but this was done
for run environments instead of build environments.

* Allow nektar to pick up system installations of openblas

* Add package for NESO

* Set up views of nektar with findable cmake configs

For some reason, cmake can't seem to find the config files for
nektar++ where it defaults to putting them. I'm not sure why, as the
documentation suggests it should be able to. In any case, in a spack
build environment this can be worked around by modifying
CMAKE_PREFIX_PATH. This option is not available for views, so instead
I added a symlink where cmake can find it.

* Fix cmake prefix path in the run environment

* Wrote package to represent DPC++ SYCL implementation

* Updated NESO package to use DPC++ one

Also fixed handling of address sanitizers

* Fix missing parameter in method declaration

* Final tidying

* Fixed typo for setting Nektar to use MKL

* Expanded documentation to reflect new capabilities

* Fix attempt to use DPC++ with GCC when building NESO

* indent typo

* Fixed typo in install instructions

* Relaxed Boost dependency requirement

Old value was from a typo in documentation

* this is the right place for the I_MPI_FABRICS warning

* add warning message about SYCL_DEVICE_FILTER

Co-authored-by: James Cook <James Cook>