Build nektar in standard Spack locations
For some reason, the Nektar++ source was being copied to a directory
in the install-tree and built there. This meant the installation was
unnecessarily large (as it contained build artefacts) and also caused
problems when trying to build Nektar++ using the development
workflow. For that, CMake was being run in the top-level directory of
the repository, before the copying took place (for some reason). This
caused problems when it was rerun with a different
compiler/toolchain.