Minor changes relating to pull request
Restored original dependencies, added conflict statements for nvcxx in neso and neso-particles.
Hipsycl temporarily has a tar.gz file and sha256sum for v0.9.4 instead of git cloning a commit.
This is since the GPU node on CSD3 doesn't seem to have functioning git cloning.
Also cuda and nvhpc packages are not necessary with the latest stable spack version (0.21.0).