Require MKL to be the same version as OneAPI compilers
This issue had been known for some time, but initially it was assumed
to be an issue with a particular release of MKL/OneAPI. However, it
appears that using a different release of MKL than of the OneAPI
compiler will always result in link-time errors. This PR adds logic to
ensure these versions always match.